// config/wallet-institution-catalog.generated.ts
export const GENERATED_EWALLET_INSTITUTIONS: WalletInstitutionEntry[] = [
  {
    "name": "Astra Pay",
    "slug": "astra-pay"
  },
  {
    "name": "Bluepay",
    "slug": "bluepay"
  },
  {
    "name": "DANA",
    "slug": "dana"
  },
  {
    "name": "Dipay",
    "slug": "dipay"
  },
  {
    "name": "DOKU",
    "slug": "doku"
  },
  {
    "name": "Dutamoney",
    "slug": "dutamoney"
  },
  {
    "name": "GoPay",
    "slug": "gopay"
  },
  {
    "name": "i.saku",
    "slug": "i-saku"
  },
  {
    "name": "JakOne Pay",
    "slug": "jakone-pay"
  },
  {
    "name": "Kaspro",
    "slug": "kaspro"
  },
  {
    "name": "LinkAja",
    "slug": "linkaja"
  },
  {
    "name": "Motion Pay",
    "slug": "motion-pay"
  },
  {
    "name": "Netzme",
    "slug": "netzme"
  },
  {
    "name": "OTTO Pay",
    "slug": "otto-pay"
  },
  {
    "name": "OVO",
    "slug": "ovo"
  },
  {
    "name": "Paydia",
    "slug": "paydia"
  },
  {
    "name": "Paytren",
    "slug": "paytren"
  },
  {
    "name": "Pospay",
    "slug": "pospay"
  },
  {
    "name": "Shopee Pay",
    "slug": "shopee-pay"
  },
  {
    "name": "SpeedCash",
    "slug": "speedcash"
  },
  {
    "name": "TrueMoney",
    "slug": "truemoney"
  },
  {
    "name": "Uangku",
    "slug": "uangku"
  },
  {
    "name": "Yukk",
    "slug": "yukk"
  }
];

export const GENERATED_INTERNATIONAL_INSTITUTIONS: WalletInstitutionEntry[] = [
  {
    "name": "Alipay",
    "slug": "alipay"
  },
  {
    "name": "Apple Pay",
    "slug": "apple-pay"
  },
  {
    "name": "BYOND by BSI",
    "slug": "byond-bsi"
  },
  {
    "name": "Flip",
    "slug": "flip"
  },
  {
    "name": "Google Pay",
    "slug": "google-pay"
  },
  {
    "name": "Grab Pay",
    "slug": "grab-pay"
  },
  {
    "name": "MoneyGram",
    "slug": "moneygram"
  },
  {
    "name": "Payoneer",
    "slug": "payoneer"
  },
  {
    "name": "PayPal",
    "slug": "paypal"
  },
  {
    "name": "Ria Money",
    "slug": "ria-money"
  },
  {
    "name": "Samsung Pay",
    "slug": "samsung-pay"
  },
  {
    "name": "Skrill",
    "slug": "skrill"
  },
  {
    "name": "SWIFT",
    "slug": "swift"
  },
  {
    "name": "Topremit",
    "slug": "topremit"
  },
  {
    "name": "Transfez",
    "slug": "transfez"
  },
  {
    "name": "WeChat Pay",
    "slug": "wechat-pay"
  },
  {
    "name": "Western Union",
    "slug": "western-union"
  },
  {
    "name": "Wise",
    "slug": "wise"
  }
];